+++ Cargo.toml
@@ -21,6 +21,7 @@ members = [
   "finchers-core",
   "finchers-derive",
   "finchers-endpoint",
+  "finchers-http",
   "finchers-json",
   "finchers-runtime",
   "finchers-urlencoded",
@@ -34,6 +35,7 @@ members = [
 finchers-core = { version = "0.11.0-dev", path = "finchers-core" }
 finchers-derive = { version = "0.11.0-dev", path = "finchers-derive" }
 finchers-endpoint = { version = "0.11.0-dev", path = "finchers-endpoint" }
+finchers-http = { version = "0.11.0-dev", path = "finchers-http" }
 finchers-runtime = { version = "0.11.0-dev", path = "finchers-runtime" }
 finchers-test = { version = "0.11.0-dev", path = "finchers-test" }
