+++ finchers-endpoint/tests/or.rs
@@ -1,12 +1,14 @@
 extern crate finchers_endpoint;
+extern crate finchers_http;
 extern crate finchers_test;
 
-use finchers_endpoint::{endpoint, ok, EndpointExt};
+use finchers_endpoint::{ok, EndpointExt};
+use finchers_http::path::path;
 use finchers_test::Client;
 
 #[test]
 fn test_or_1() {
-    let endpoint = endpoint("foo").right(ok("foo")).or(endpoint("bar").right(ok("bar")));
+    let endpoint = path("foo").right(ok("foo")).or(path("bar").right(ok("bar")));
     let client = Client::new(endpoint);
 
     let outcome = client.get("/foo").run().unwrap();
@@ -18,8 +20,8 @@ fn test_or_1() {
 
 #[test]
 fn test_or_choose_longer_segments() {
-    let e1 = endpoint("foo").right(ok("foo"));
-    let e2 = endpoint("foo/bar").right(ok("foobar"));
+    let e1 = path("foo").right(ok("foo"));
+    let e2 = path("foo/bar").right(ok("foobar"));
     let endpoint = e1.or(e2);
     let client = Client::new(endpoint);
