+++ finchers-http/src/body.rs
@@ -17,11 +17,11 @@
 use finchers_core::error::BadRequest;
 use finchers_core::input;
 use finchers_core::{Bytes, BytesString, Input, Never};
+use finchers_endpoint::{Context, Endpoint, Error};
 use futures::{Future, Poll};
 use std::marker::PhantomData;
 use std::str::Utf8Error;
 use std::{error, fmt};
-use {Context, Endpoint, Error};
 
 /// Creates an endpoint for parsing the incoming request body into the value of `T`
 pub fn body<T: FromBody>() -> Body<T> {
