+++ finchers-http/src/lib.rs
@@ -1,0 +1,14 @@
+extern crate finchers_core;
+extern crate finchers_endpoint;
+extern crate http;
+#[macro_use]
+extern crate futures;
+
+pub mod body;
+pub mod header;
+pub mod method;
+pub mod path;
+
+pub use body::FromBody;
+pub use header::FromHeader;
+pub use path::{FromSegment, FromSegments};
