+++ finchers-json/Cargo.toml
@@ -14,8 +14,7 @@ maintenance = { status = "actively-developed" }
 
 [dependencies]
 finchers-core = { version = "0.11.0-dev", path = "../finchers-core" }
-finchers-endpoint = { version = "0.11.0-dev", path = "../finchers-endpoint" }
-futures = "0.1.20"
+finchers-http = { version = "0.11.0-dev", path = "../finchers-http" }
 http = "0.1.5"
 mime = "0.3.0"
 serde = "1.0"
