+++ src/lib.rs
@@ -3,6 +3,7 @@ extern crate finchers_core;
 #[macro_use]
 extern crate finchers_derive;
 extern crate finchers_endpoint;
+extern crate finchers_http;
 extern crate finchers_json;
 extern crate finchers_runtime;
 extern crate finchers_urlencoded;
@@ -17,16 +18,17 @@ pub use finchers_derive::*;
 pub use finchers_core::{error, output};
 
 pub mod endpoint {
-    pub use finchers_endpoint::{all, body, endpoint, header, method, ok, path, skip_all, Endpoint, EndpointExt,
-                                FromBody, FromHeader, FromSegment, FromSegments, IntoEndpoint};
+    pub use finchers_endpoint::{all, endpoint, ok, skip_all, Endpoint, EndpointExt, IntoEndpoint};
+
+    pub use finchers_http::{body, header, method, path, FromBody, FromHeader, FromSegment, FromSegments};
 
     /// The "prelude" for building endpoints
     pub mod prelude {
-        pub use finchers_endpoint::body::{body, body_stream};
-        pub use finchers_endpoint::header::{header, header_opt, header_req};
-        pub use finchers_endpoint::method::{delete, get, head, patch, post, put};
-        pub use finchers_endpoint::path::{match_, path, paths};
         pub use finchers_endpoint::{endpoint, Endpoint, IntoEndpoint};
+        pub use finchers_http::body::{body, body_stream};
+        pub use finchers_http::header::{header, header_opt, header_req};
+        pub use finchers_http::method::{delete, get, head, patch, post, put};
+        pub use finchers_http::path::{param, params, path};
     }
 }
 
@@ -39,12 +41,12 @@ pub mod runtime {
 }
 
 pub mod json {
-    pub use finchers_json::{json_body, Error, Json, JsonBody, JsonOutput};
+    pub use finchers_json::{Error, Json};
 }
 
 pub mod urlencoded {
-    pub use finchers_urlencoded::{form_body, from_csv, queries, queries_opt, queries_req, Error, Form, FormBody,
-                                  Queries, QueriesOptional, QueriesRequired};
+    pub use finchers_urlencoded::{from_csv, queries, queries_opt, queries_req, Error, Form, Queries, QueriesOptional,
+                                  QueriesRequired};
 }
 
 pub mod prelude {
